+++ torchmetrics/__init__.py
@@ -55,7 +55,13 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
-from torchmetrics.image import PSNR, SSIM, MultiScaleStructuralSimilarityIndexMeasure  # noqa: E402
+from torchmetrics.image import (  # noqa: E402
+    PSNR,
+    SSIM,
+    MultiScaleStructuralSimilarityIndexMeasure,
+    PeakSignalNoiseRatio,
+    StructuralSimilarityIndexMeasure,
+)
 from torchmetrics.metric import Metric  # noqa: E402
 from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
@@ -150,6 +156,7 @@ __all__ = [
     "Precision",
     "PrecisionRecallCurve",
     "PSNR",
+    "PeakSignalNoiseRatio",
     "R2Score",
     "Recall",
     "RetrievalFallOut",
@@ -175,6 +182,7 @@ __all__ = [
     "Specificity",
     "SQuAD",
     "SSIM",
+    "StructuralSimilarityIndexMeasure",
     "StatScores",
     "SumMetric",
     "SymmetricMeanAbsolutePercentageError",