+++ torchmetrics/functional/audio/si_snr.py
@@ -11,13 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from warnings import warn
 
+from deprecate import deprecated, void
 from torch import Tensor
 
 from torchmetrics.functional.audio.snr import scale_invariant_signal_noise_ratio
 
 
+@deprecated(target=scale_invariant_signal_noise_ratio, deprecated_in="0.7", remove_in="0.8")
 def si_snr(preds: Tensor, target: Tensor) -> Tensor:
     """Scale-invariant signal-to-noise ratio (SI-SNR).
 
@@ -29,8 +30,4 @@ def si_snr(preds: Tensor, target: Tensor) -> Tensor:
         >>> si_snr(torch.tensor([2.5, 0.0, 2.0, 8.0]), torch.tensor([3.0, -0.5, 2.0, 7.0]))
         tensor(15.0918)
     """
-    warn(
-        "`si_snr` was renamed to `scale_invariant_signal_noise_ratio` in v0.7 and it will be removed in v0.8",
-        DeprecationWarning,
-    )
-    return scale_invariant_signal_noise_ratio(preds, target)
+    return void(preds, target)