+++ torchmetrics/functional/text/wer.py
@@ -13,10 +13,9 @@
 # limitations under the License.
 
 from typing import List, Tuple, Union
-from warnings import warn
 
 import torch
-from deprecate import deprecated
+from deprecate import deprecated, void
 from torch import Tensor, tensor
 
 from torchmetrics.functional.text.helper import _edit_distance
@@ -98,11 +97,10 @@ def word_error_rate(preds: Union[str, List[str]], target: Union[str, List[str]])
     return _wer_compute(errors, total)
 
 
-def wer(  # ToDo: remove in v0.8
-    preds: Union[None, str, List[str]] = None,
-    target: Union[None, str, List[str]] = None,
-    predictions: Union[None, str, List[str]] = None,
-    references: Union[None, str, List[str]] = None,
+@deprecated(target=word_error_rate, deprecated_in="0.7", remove_in="0.8")
+def wer(
+    predictions: Union[str, List[str]],
+    references: Union[str, List[str]],
 ) -> Tensor:
     """Word error rate (WER_) is a common metric of the performance of an automatic speech recognition system.
 
@@ -115,24 +113,4 @@ def wer(  # ToDo: remove in v0.8
         >>> wer(preds=preds, target=target)
         tensor(0.5000)
     """
-    assert preds is not None or predictions is not None
-    assert target is not None or references is not None
-
-    if predictions is not None:
-        warn(
-            "You are using deprecated argument `predictions` in v0.7 which was renamed to `preds`. "
-            " The past argument will be removed in v0.8.",
-            DeprecationWarning,
-        )
-        preds = predictions
-
-    if references is not None:
-        warn(
-            "You are using deprecated argument `references` in v0.7 which was renamed to `target`. "
-            " The past argument will be removed in v0.8.",
-            DeprecationWarning,
-        )
-        target = references
-
-    warn("`wer` was renamed to `word_error_rate` in v0.7 and it will be removed in v0.8", DeprecationWarning)
-    return word_error_rate(preds, target)
+    return void(predictions, references)